Fix transparent search box when using Unity free
Since the box texture is only set when you happen to be using the pro
version of Unity, the search box would be transparent in Unity free.

Since the Unity free box texture is not partially transparent, it is
not necessary to add logic similar to the pro version of the code.